make

+++ Makefile
@@ -41,9 +41,7 @@ mysql:
 
 .PHONY: preparekub
 preparekub:
-	if [ ! -d "$HOME/google-cloud-sdk/bin" ]; then
-	rm -rf "$HOME/google-cloud-sdk"
-	curl https://sdk.cloud.google.com | bash > /dev/null
+	if [ ! -d "$HOME/google-cloud-sdk/bin" ]; then rm -rf "$HOME/google-cloud-sdk" && curl https://sdk.cloud.google.com | bash > /dev/null
 	fi
 	# Promote gcloud to PATH top priority (prevent using old version fromt travis)
 	source $HOME/google-cloud-sdk/path.bash.inc
